Funzioni di libreria di Gui per terminale completate. 'gui.c' è un file di demo
